Move non-page files to components folder and update imports
