+++ app/main/index.tsx
@@ -4,7 +4,7 @@ import { View, Text, Button, FlatList, TouchableOpacity, TextInput, StyleSheet }
 import { collection, query, where, onSnapshot, addDoc, Timestamp, doc, getDoc } from 'firebase/firestore';
 import { auth, db } from '../../firebase';  // Firebase setup
 import { useTheme } from 'react-native-paper';
-import {} from '../PaperThemes';
+import {} from '../../components/PaperThemes';
 import { Redirect, useRouter } from 'expo-router';
 
 export default function Events() {
